+++ demos/chinook/src/main/java/is/codion/demos/chinook/model/ChinookAppModel.java
@@ -22,7 +22,7 @@ import is.codion.common.version.Version;
 import is.codion.demos.chinook.domain.api.Chinook.Customer;
 import is.codion.demos.chinook.domain.api.Chinook.PlaylistTrack;
 import is.codion.framework.db.EntityConnectionProvider;
-import is.codion.framework.model.ForeignKeyDetailModelLink;
+import is.codion.framework.model.ForeignKeyModelLink;
 import is.codion.swing.framework.model.SwingEntityApplicationModel;
 import is.codion.swing.framework.model.SwingEntityModel;
 
@@ -48,7 +48,7 @@ public final class ChinookAppModel extends SwingEntityApplicationModel {
 		SwingEntityModel playlistModel = new SwingEntityModel(new PlaylistTableModel(connectionProvider));
 		SwingEntityModel playlistTrackModel = new SwingEntityModel(new PlaylistTrackEditModel(connectionProvider));
 
-		playlistModel.detailModels().add(ForeignKeyDetailModelLink.builder(playlistTrackModel, PlaylistTrack.PLAYLIST_FK)
+		playlistModel.detailModels().add(ForeignKeyModelLink.builder(playlistTrackModel, PlaylistTrack.PLAYLIST_FK)
 						.clearValueOnEmptySelection(true)
 						.active(true)
 						.build());