+++ changelog.md
@@ -35,6 +35,7 @@ Codion Change Log
 - DefaultEntityModel, type parameters now refer to EntityModel and EntityEditModel instead of DefaultEntityModel and AbstractEntityEditModel.
 - DefaultEntityModel now abstract, renamed AbstractEntityModel.
 - DetailModelLink and ForeignKeyDetailModelLink now Builder based, implementations package private and final.
+- DetailModelLink and ForeignKeyDetailModelLink renamed ModelLink and ForeignKeyModelLink.
 ### is.codion.swing.framework.model
 - DefaultSwingForeignKeyConditionModel now links search and combo box models to their respective operands.
 ### is.codion.swing.framework.ui